requirements.txt
agno==1.8.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
backoff==2.2.1
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.7
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
emoji==2.14.1
fastembed==0.7.3
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.32.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
joblib==1.5.2
langdetect==1.0.9
loguru==0.7.3
lxml==6.0.1
Markdown==3.8.2
markdown-it-py==4.0.0
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
mypy_extensions==1.1.0
nltk==3.9.1
numpy==2.3.2
olefile==0.47
ollama==0.5.3
onnxruntime==1.22.1
openai==1.102.0
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.0
psutil==7.0.0
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.0.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.2
qdrant-client==1.15.1
RapidFuzz==3.14.0
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.0
tomli==2.2.1
tqdm==4.67.1
typer==0.17.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
unstructured==0.18.14
unstructured-client==0.42.3
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
